[abucoins] Adding return status of 'closed'.
It seems there's a return type of closed returned for status as well.  It's listed in some examples, but not in the textual part of the API description.  I've reached out to Abucoins for clarity on what it means, but the best mapping I could think of was PENDING_CANCEL.